backend/requirements.txt
# Backend Python Dependencies
# TB Personal OS - Assistant API

# Framework
fastapi[all]==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Supabase
supabase>=2.3.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-auth>=2.26.0  # Versão compatível com generativeai

# AI/LLM
google-generativeai>=0.8.0
openai>=1.3.0

# Telegram
python-telegram-bot==13.15

# HTTP Client
aiohttp>=3.9.0

# Data Processing
pandas==2.0.3
numpy==1.24.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3
bleach==6.1.0
python-dotenv==1.0.0

# Scheduler (versão compatível com python-telegram-bot)
APScheduler==3.6.3

# Logging & Monitoring
structlog>=24.0.0
python-json-logger==2.0.7
sentry-sdk==1.39.0

# Rate Limiting
slowapi==0.1.9

# HTTP Requests
requests>=2.31.0
httpx>=0.25.0

# Async Redis
redis>=5.0.0

# Validation
email-validator==2.1.0
phonenumbers==8.13.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx>=0.25.0
faker==20.1.0

# Development
black==23.12.0
isort==5.13.0
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0

# Production
gunicorn==21.2.0

# Monitoring & Observability
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.0
